Move dictionary service to services folder
