+++ src/app/service/dictionaries/dictionary.service.ts
@@ -1,5 +1,5 @@
-import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
 import { environment as env } from 'environments/environment';
 
 @Injectable({
